[feat] add master user resources

+++ resources/js/lib/local_storage.ts
@@ -1,7 +1,16 @@
 export const saveLocalStorage = (key: string, value: any) => {
     if (typeof window !== "undefined") {
+        if (existLocalStorage(key)) return;
         window.localStorage.setItem(key, JSON.stringify(value));
+        window.location.reload();
     }
+};
+
+export const existLocalStorage = (key: string) => {
+    if (typeof window !== "undefined") {
+        return window.localStorage.getItem(key) !== null;
+    }
+    return false;
 };
 
 export const getLocalStorage = (key: string) => {
